feat: 策略模式优惠折扣计算

+++ group-buy-market-infrastructure/src/main/java/cn/bugstack/infrastructure/dao/po/GroupBuyDiscount.java
@@ -1,5 +1,6 @@
 package cn.bugstack.infrastructure.dao.po;
 
+import cn.bugstack.domain.activity.model.valobj.DiscountTypeEnum;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -27,7 +28,7 @@ public class GroupBuyDiscount {
     private String discountDesc;
 
     /* 折扣类型（0:base、1:tag） */
-    private Byte discountType;
+    private Integer discountType;
 
     /* 营销优惠计划（ZJ:直减、MJ:满减、N元购） */
     private String marketPlan;
